Filter out colors from the list
Removed the colors that came from the background, as we wanted only the
dot colors, and saved the filtered list to color_list variable.

Commented out the rest of the code as we don't need to run it everytime
we run the code.

Our job to extract out the main colors is done at color_list.